Create lightcurves.py
Exported Jupyter notebook to .py, then uploaded to Github. There are things in the code which may not be clear, and there is a section at the end that is not yet working, but I am uploading this now in case you would like to see my code over the weekend. I will endeavour to improve the code either tomorrow or on Monday.