Code for running simulations
This code runs simulations for various input parameters and fits them with various specified fitting parameters in the usual SNTD way. It then pickles the relevant lightcurve input and fitted parameters so that analysis can be done in future without rerunning the simulations, in separate scripts. A problem with it was that it created data and fitted it in the same loop, and because of the issue with systematic random scaling of the generated lightcurve, the results can be sensitive to variable normalisations of the lightcurves within the loop. The shapes of the lightcurves (and thus the fitted time delays) stay the same, so some of the data, as I understand, should still be useful.