Create fitting_pick.py
Added an example script that takes stored pickle file from resolved_llc.py and unpacks some of the relevant data and prints it for sake of example. The last cell in the script unpickles the magnitudes vs Av simulation file, extracts and formats the information contained in a useful way, and plots the ratios for various extinctions, showing how the pickle files can be used for further analysis.