Update lightcurve generation.py
This code just generates lightcurve data which can then be imported and used in analysis. This was used to get around the systematic normalisation changes when creating the data within a simulation. This just makes sure that in future analysis, the lightcurves used always have the same normalisation - in other words, to ensure no funny business!