Create priors_testing.py
This is the code I have been using to test the effect of priors/bounds on x1 and c on the time delay fitting and the errors. The relevant bounds can be modified in I hope a simple manner, as can the parameters being fitted. This code makes use of the light curve generator code to ensure all lc's have same normalisation. Hopefully the code is fairly straightforward but please let me know if anything is unclear :)